i386: fw: Add SMBIOS and MSR feature control to fw_cfg
We expand the fw_cfg device data with SMBIOS properties and a
description of the MSR feature control we want to expose to the
firmware.
OVMF will need this in order to properly initialize and boot a guest
Linux kernel.
